+++ docs/conf.py
@@ -9,6 +9,7 @@ import nengo_bones
 extensions = [
     "sphinx.ext.autodoc",
     "sphinx.ext.autosummary",
+    "sphinx.ext.doctest",
     "sphinx.ext.githubpages",
     "sphinx.ext.intersphinx",
     "sphinx.ext.mathjax",
@@ -24,6 +25,13 @@ extensions = [
 autoclass_content = "both"  # class and __init__ docstrings are concatenated
 autodoc_default_options = {"members": None}
 autodoc_member_order = "bysource"  # default is alphabetical
+
+# -- sphinx.ext.doctest
+doctest_global_setup = """
+import nengo_bones
+# Testing that doctest_setup works
+# with multiple lines
+"""
 
 # -- sphinx.ext.intersphinx
 intersphinx_mapping = {